Create tests for bugs reported about empty lines in strings vs indentation (#159)
Those are all of the examples that match the reported indentation
errors.

The very strange thing is that all of the new tests fail, but they fail
differently in vader than how they do in an interactive (n)vim session.
The result that we get in those tests is usually that the string
contents get wrongly moved to fit indentation. However, when I run the
indentation command in the editor, the string contents don't change but
things afterwards get indented at the wrong level. I'm not sure why the
tests don't behave in the same way.